Feature/korchagin skill calc

+++ site.local/database/migrations/2020_07_04_135645_init.php
@@ -204,7 +204,7 @@ class Init extends Migration
     protected function createStudentSkill()
     {
         Schema::create('student_skill', function (Blueprint $table) {
-            $table->integer('points');
+            $table->float('points');
             $table->integer('student_id');
             $table->integer('skill_id');
             $table->unique(['student_id', 'skill_id']);
